backend/requirements.txt
Flask==3.0.3
gunicorn==23.0.0
psycopg2-binary==2.9.9
Werkzeug==3.0.3
flask-cors==4.0.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
email-validator==2.1.1
APScheduler==3.10.4
python-dateutil==2.9.0
Authlib==1.3.1
requests==2.32.3
gevent==24.2.1
setuptools<81
apprise==1.9.3
